Move view order and confirmation into Order namespace
